mood structure update
