+++ daily-mood/daily-mood.vcxproj
@@ -119,6 +119,7 @@
     <ClInclude Include="fileReader.h" />
     <ClInclude Include="helpers.h" />
     <QtMoc Include="TodoCard.h" />
+    <ClInclude Include="mood.h" />
     <ClInclude Include="todo.h" />
     <ClInclude Include="ui_eventdialog.h" />
   </ItemGroup>
