+++ daily-mood/daily-mood.vcxproj.filters
@@ -99,6 +99,9 @@
     <ClInclude Include="todo.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="mood.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <QtUic Include="EventDialog.ui">
