+++ daily-mood/x64/Debug/daily-mood.log
@@ -1,3 +1,4 @@
 ﻿  appData.cpp
   main.cpp
+  mood.cpp
   daily-mood.vcxproj -> C:\Users\poczt\Documents\school\Studia-I-politechnika\JiPP2\projekt-koncowy\x64\Debug\daily-mood.exe
